MARKED ALL REST TESTS. COOL Configuration to swith on the Fly Provider Configuration

+++ test/geoplatform-services-test/src/test/java/org/geosdi/geoplatform/model/soap/WSFolderTest.java
@@ -260,7 +260,7 @@ public class WSFolderTest extends BaseSoapServiceTest {
 
         Assert.assertEquals("totalElementsOfProject after added",
                 totalElementsOfProject + 1,
-                gpWSClient.getNumberOfElementsProject(idProjectTest));
+                gpWSClient.getNumberOfElementsProject(idProjectTest).intValue());
 
         this.checkState(new int[]{8, 7, 6, 5, 4, 3, 2}, new int[]{2, 3},
                 "before removing");
@@ -273,7 +273,7 @@ public class WSFolderTest extends BaseSoapServiceTest {
 
         Assert.assertEquals("totalElementsOfProject after deleted",
                 totalElementsOfProject, gpWSClient.getNumberOfElementsProject(
-                        idProjectTest));
+                        idProjectTest).intValue());
 
         this.checkInitialState("after removing");
     }
@@ -309,7 +309,7 @@ public class WSFolderTest extends BaseSoapServiceTest {
 
         Assert.assertEquals("totalElementsOfProject after added",
                 totalElementsOfProject + 1,
-                gpWSClient.getNumberOfElementsProject(idProjectTest));
+                gpWSClient.getNumberOfElementsProject(idProjectTest).intValue());
 
         this.checkState(new int[]{8, 7, 6, 5, 4, 2, 1}, new int[]{2, 4},
                 "before removing");
@@ -324,7 +324,7 @@ public class WSFolderTest extends BaseSoapServiceTest {
 
         Assert.assertEquals("totalElementsOfProject after deleted",
                 totalElementsOfProject, gpWSClient.getNumberOfElementsProject(
-                        idProjectTest));
+                        idProjectTest).intValue());
 
         this.checkInitialState("after removing");
     }
